requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
streamlit>=1.40.0
pydantic>=2.9.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
plotly>=5.24.0
